+++ package-testing/spec/package/version_selection_spec.rb
@@ -33,13 +33,13 @@ describe 'Test puppet & ruby version selection' do
       describe command('pdk bundle exec puppet --version') do
         its(:exit_status) { is_expected.to eq(0) }
         its(:stderr) { is_expected.to match(%r{using puppet (#{expected_puppets})}im) }
-        its(:stderr) { is_expected.to match(%r{^(#{expected_puppets})$}im) }
+        its(:stdout) { is_expected.to match(%r{^(#{expected_puppets})$}im) }
       end
 
       describe command('pdk bundle exec ruby --version') do
         its(:exit_status) { is_expected.to eq(0) }
         its(:stderr) { is_expected.to match(%r{using ruby #{Regexp.escape(test_case[:expected_ruby])}[\.0-9]*}im) }
-        its(:stderr) { is_expected.to match(%r{^ruby #{Regexp.escape(test_case[:expected_ruby])}[\.0-9]*p}im) }
+        its(:stdout) { is_expected.to match(%r{^ruby #{Regexp.escape(test_case[:expected_ruby])}[\.0-9]*p}im) }
       end
     end
   end
